Migrate fjs/types/rtti/ts to .f.mjs, split types into types.ts
Splits the full Ts<T> type-transformer API (Primitive, Unknown, Array,
Object, Info0Ts, ConstTs, Info1Ts, ArrayTs, RecordTs, TupleTs, StructTs,
Ts) into a sibling types.ts, and converts the sole runtime export
(printer) to JSDoc-typed module.f.mjs/proof.f.mjs under the stage-1
TypeScript-to-mjs migration.

Updates 11 dependents (protocol/mcp, protocol/json_rpc,
dev/package_json, ci/common, media/json, media/json/schema,
media/revision, media, emergent_testing, and rtti's own
common/parse/validate submodules) to import the Ts type from types.ts.
All were type-only imports of Ts/Unknown, so no runtime specifier
changes were needed beyond the import path.

Fixes two stale doc references to fjs/types/rtti/ts/module.f.ts and
proof.f.ts.
