+++ fjs/media/module.f.ts
@@ -47,7 +47,7 @@ import { detectVec, type DetectMeta } from './type/module.f.ts'
 import { parse } from './json/module.f.ts'
 import { assert, assertNotNullish } from '../asserts/module.f.mjs'
 import type { Struct } from '../types/rtti/types.ts'
-import type { Ts, Unknown } from '../types/rtti/ts/module.f.ts'
+import type { Ts, Unknown } from '../types/rtti/ts/types.ts'
 import { validate, type Validate } from '../types/rtti/validate/module.f.ts'
 
 /**
