+++ fjs/protocol/json_rpc/module.f.ts
@@ -15,7 +15,7 @@
  * @module
  */
 import { number, string, or, option } from '../../types/rtti/module.f.mjs'
-import type { Ts } from '../../types/rtti/ts/module.f.ts'
+import type { Ts } from '../../types/rtti/ts/types.ts'
 import { validate } from '../../types/rtti/validate/module.f.ts'
 import type { Result } from '../../types/result/types.ts'
 import { unknown, type Unknown } from '../../media/json/module.f.ts'
