+++ fjs/types/rtti/parse/proof.f.ts
@@ -1,7 +1,7 @@
 import { parse, type ValidationError } from './module.f.ts'
 import { boolean, number, string, bigint, unknown, array, record, or, option } from '../module.f.mjs'
 import type { Equal } from '../../ts/types.ts'
-import type { Ts } from '../ts/module.f.ts'
+import type { Ts } from '../ts/types.ts'
 import type { Unknown as DjsUnknown } from '../../../djs/module.f.ts'
 import type { Assert } from '../../../asserts/types.ts'
 import { assert, assertEq } from '../../../asserts/module.f.mjs'
