+++ fjs/types/rtti/todo/662.md
@@ -49,11 +49,11 @@ const visitConst = <R>(v: Visitor<R>) => (c: Const): R =>
 `validate` and `parse` both compose their top-level function from a
 `Visitor<R>` and call `visit(...)`. But there is a **third** walker over
 the same `Type` ADT in the same `rtti/` tree — the runtime `printer` in
-`fjs/types/rtti/ts/module.f.ts` — and it never adopted `visit`. It
+`fjs/types/rtti/ts/module.f.mjs` — and it never adopted `visit`. It
 re-hand-rolls the identical recognition logic:
 
 ```ts
-// fjs/types/rtti/ts/module.f.ts:118
+// fjs/types/rtti/ts/module.f.mjs:118
 export const printer = (mut?: true): (rtti: Type) => string => {
     const { tuple, struct, array, record } = tsPrinter(mut)
 
@@ -214,7 +214,7 @@ same module (`:44`, `:55`).
 
 - `fjs/types/rtti/common/module.f.ts:99,124` — `visitConst` / `visit`, the
   shared schema walker.
-- `fjs/types/rtti/ts/module.f.ts:118-145` — the printer's duplicated
+- `fjs/types/rtti/ts/module.f.mjs:118-145` — the printer's duplicated
   `constToTs` + `toTs` switch.
 - `fjs/types/ts/module.f.mjs:23-29,44,55` — the printer leaves the visitor
   handlers delegate to.
